% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSeq_fit.R
\name{corrSeq_fit}
\alias{corrSeq_fit}
\title{Function to fit various types of models to longitudinal RNA-Seq data}
\usage{
corrSeq_fit(
  formula = NULL,
  expr_mat = NULL,
  gene_names = NULL,
  sample_data = NULL,
  method,
  id,
  small.samp.method = NULL,
  parallel = F,
  cores = 2,
  ...
)
}
\arguments{
\item{formula}{A one-sided linear formula describing the model variables. For models that include them, random effects should be included in the formula using the syntax of the lme4 package.}

\item{expr_mat}{A (G x N) numeric matrix RNA-seq expression data with genes in rows and samples in columns. For \code{method="gee"}, \code{method="nbmm_pl"}, \code{method="nbmm_lp"} and \code{method="nbmm_agq"}, the matrix should contain raw counts and for \code{method="lmm"} the matrix should contain transformed counts (e.g. using VST from DESeq2). G = number of genes.  N = number of samples.}

\item{gene_names}{An optional character vector of gene names (length G).}

\item{sample_data}{Data frame with N rows containing the fixed- and random-effects terms included in the formula.  The rows of the data frame must correspond (and be in the same order as) the columns of the expression matrix.}

\item{method}{Method to use to fit the models. Possible options are \code{"lmm"}, \code{"gee"}, \code{"nbmm_pl"}, \code{"nbmm_lp"} and \code{"nbmm_agq"}.}

\item{id}{Only applicable for models fit using \code{method="gee"}. A vector or data column name which identifies the clusters. The length of
‘id’ should be the same as the number of observations. Data are
assumed to be sorted so that observations on each cluster appear
as contiguous rows in data. If data is not sorted this way, the
function will not identify the clusters correctly. If \code{sort=TRUE} (default),
the dataframe from the \code{data} argument is sorted by the id column to avoid
this issue.}

\item{small.samp.method}{Only applicable for models fit using \code{method="gee"}. A character string specifying the
small sample method. The following are permitted: "pan" for the
Pan (2001) method, "md" for the Mancl and Derouen (2001) method, and "wl" for the Wang and Long (2011) method.
If \code{small.samp.method} is null, small sample variance estimates are not computed.}

\item{parallel}{A logical variable indicating whether forking (via mclapply) should be used to parallelize fits. Only available on Mac or linux machines.}

\item{cores}{Number of cores to use if parallelizing (default is 2)}

\item{...}{additional arguments passed on to \code{\link[lmerTest]{lmer}} (\code{method="lmm"}), \code{\link{gee_small_sample}} (\code{method="gee"}),
\code{\link{glmm_nb_lmer}} (\code{method="nbmm_pl"}), \code{\link[glmmADMB]{glmmadmb}} (\code{method="nbmm_lp"}), or \code{\link[GLMMadaptive]{mixed_model}} (method="nbmm_agq").}
}
\value{
A list of length G of model objects from the following functions: \code{\link[lmerTest]{lmer}} (\code{method="lmm"}), \code{\link{gee_small_sample}}(\code{method="lmm"}),
\code{\link{glmm_nb_lmer}} (\code{method="nbmm_pl"}), \code{\link[glmmADMB]{glmmadmb}} (\code{method="nbmm_lp"}), or \code{\link[GLMMadaptive]{mixed_model}} (method="nbmm_agq").
}
\description{
Wrapper function that fits one of five types of models to RNA-Seq data. Available model fitting methods are linear mixed models (lmm) (using transformed data), generalized estimating equations with an optional small sample adjustment (gee), and negative binomial models using
a pseudo-likelhood approach (nbmm_pl), a maximum likelihood approach with Laplace approximation (nbmm_lp), or a maximum likelihood approach with adaptive Gaussian quadrature (nbmm_agq).
}
\examples{
data("simdata")
sample_meta_data <- simdata$metadata

## Subset down to 10 observation (i.e. gene)
counts=simdata$counts[1:10,]

## Fit GEE models using Wang-Long small sample size estimator
## log_offset is the log size factors from the DESeq2 package
gee_fit <- corrSeq_fit(formula = ~ group * time+offset(log_offset),
                           expr_mat = counts,
                           sample_data = sample_meta_data,
                           method="gee",
                           id=ids,
                           small.samp.method="wl")

## Fit NBMM-PL models
nbmm_pl_fit <- corrSeq_fit(formula = ~ group * time+(1|ids)+offset(log_offset),
                           expr_mat = counts,
                           sample_data = sample_meta_data,
                           method="nbmm_pl")


## Fit NBMM-AGQ models
nbmm_agq_fit <- corrSeq_fit(formula = ~ group * time+(1|ids)+offset(log_offset),
                           expr_mat = counts,
                           sample_data = sample_meta_data,
                           method="nbmm_agq")

## Fit NBMM-LP models
## Random effects must be factors
sample_meta_data$ids<-factor(sample_meta_data$ids)
nbmm_lp_fit <- corrSeq_fit(formula = ~ group * time+(1|ids)+offset(log_offset),
                           expr_mat = counts,
                           sample_data = sample_meta_data,
                           method="nbmm_lp")

## Fit LMM models to transformed data
## Use the variance transformed counts in the simdata object
## Subset down to 10 genes
vst_expr<-simdata$vst_expr[1:10,]

##  Fit the Models
lmm_fit<- corrSeq_fit(formula = ~ group * time + (1|ids),
                           expr_mat = vst_expr,
                           sample_data = sample_meta_data,
                           method="lmm")
}
\seealso{
\code{\link{corrSeq_summary}}, \code{\link{glmm_nb_lmer}}, \code{\link[lmerTest]{lmer}}, \code{\link{gee_small_sample}}, \code{\link[glmmADMB]{glmmadmb}}, and \code{\link[GLMMadaptive]{mixed_model}}
}
\author{
Elizabeth Wynn
}
