% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSeq_summary.R
\name{corrSeq_summary}
\alias{corrSeq_summary}
\title{Function to summarize individual regression coefficients}
\usage{
corrSeq_summary(
  corrSeq_results = NULL,
  coefficient = NULL,
  p_adj_method = "BH",
  ddf = "Satterthwaite",
  sort_results = T
)
}
\arguments{
\item{corrSeq_results}{Results object from running corrSeq_fit}

\item{coefficient}{Character string or numeric indicator of which coefficient to summarize}

\item{p_adj_method}{Method for adjusting for multiple comparisons (default is Benjamini-Hochberg). See p.adjust.methods}

\item{ddf}{Method for computing degrees of freedom and t-statistics.
The options "Satterthwaite" and "Kenward-Roger" can be used for
models fit only using nbmm_pl or lmm. Options "containment" and
"residual" can be used for models fit using any method. Alternatively, a single numeric value representing the ddf for all tests can also be given.}

\item{sort_results}{Should the results table be sorted by adjusted p-value?}
}
\description{
Conducts t-tests on individual regression coefficients from fits done using lmerSeq_fit function
}
\examples{
data("simdata")
sample_meta_data <- simdata$metadata

## Subset down to 10 observation (i.e. gene)
counts=simdata$counts[1:10,]

## Fit NBMM-PL models
nbmm_pl_fit <- corrSeq_fit(formula = ~ group * time+(1|ids),
                          expr_mat = counts,
                          sample_data = sample_meta_data,
                          method="nbmm_pl")

## Summarize the group coefficient with Satterthwaite degrees of freedom
model_sum <-corrSeq_summary(corrSeq_results = nbmm_pl_fit,
                             coefficient = "group",
                             p_adj_method = 'BH',
                             ddf = 'Satterthwaite',
                             sort_results = T)


}
