% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_small_sample.R
\name{geeglm_small_samp}
\alias{geeglm_small_samp}
\title{Fit Generalize Estimating Equations with Small Sample Variance Estimators}
\usage{
geeglm_small_samp(
  formula,
  family = poisson,
  data,
  id,
  corstr = "exchangeable",
  small.samp.method = "wl",
  sort = T,
  ...
)
}
\arguments{
\item{formula}{A two-sided linear formula object with the response
on the left of a ~ operator and the terms, separated by + operators, on the right.}

\item{family}{a description of the error distribution and link
function to be used in the model. Only \code{guassian} and \code{poisson}
families supported.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model and the id variable. If not
environment(formula), typically the environment from which the function is called.}

\item{id}{a vector or data column name which identifies the clusters. The length of
‘id’ should be the same as the number of observations. Data are
assumed to be sorted so that observations on each cluster appear
as contiguous rows in data. If data is not sorted this way, the
function will not identify the clusters correctly. If \code{sort=TRUE} (default),
the dataframe from the \code{data} argument is sorted by the id column to avoid
this issue.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: "independence", "exchangeable", "ar1", and "unstructured"}

\item{small.samp.method}{a character string specifying the
small sample method. The following are permitted: "pan" for the
Pan (2001) method, "md" for the Mancl and Derouen (2001) method, and "wl" for the Wang and Long (2011) method.
If \code{small.samp.method} is null, small sample variance estimates are not computed.
The resulting object will be identical to the object created if
\code{geeglm} from the \pkg{geepack}-package was used.}

\item{...}{additional arguments passed on to \code{geepack::geeglm}.}
}
\value{
A \code{geeglm} object with an additional item
\code{small.samp.var} which contains the small sample variance estimators using the specified method.
}
\description{
This function fits a GEE model using the \code{geeglm} function
from the \pkg{geepack}-package. Additionally, small sample variance
estimates are calculated using one of three methods proposed in the following
papers: Pan (2001), Mancl and Derouen (2001), and Wang and Long (2011).
}
\details{
This function borrows heavily from the corresponding small sample variance estimating
functions in the \pkg{geesmv}-package
(\code{\link[geesmv]{GEE.var.pan}}, \code{\link[geesmv]{GEE.var.md}} and \code{\link[geesmv]{GEE.var.wl}}).
In addition to combining these functions into a single function and using the \code{geeglm} function
in model fitting, this function also varys from these functions in that it
ensures that model offsets are properly accounted for in the calculation of
small sample estimators.
}
\examples{
data("simdata")
sample_meta_data <- simdata$metadata

#Subset down to one observation (i.e. gene)
counts=simdata$counts[1,]

#Combine counts, metadata into dataframe
df=cbind(counts, sample_meta_data)

#Sort data by id (Function also does this if sort=T)
df=df[order(df$ids),]

#Fit the Model-use Pan method for small sample variance
fit.gee.pan<-geeglm_small_samp(formula =counts ~ group * time,
                              family=poisson, data=df, id=ids,
                              corstr="exchangeable",
                              small.samp.method="pan", sort=T)
}
\references{
Mancl LA, DeRouen TA (2001). "A covariance estimator for GEE with improved small-sample properties." Biometrics, 57(1), 126-134. ISSN 0006341X. doi:10. 1111/j.0006-341X.2001.00126.x.

Pan W (2001). "On the robust variance estimator in generalised estimating equations." \emph{Biometrika}, 88(3), 901-906. ISSN 00063444. doi:10.1093/biomet/88.3.901.

Wang M, Long Q (2011). "Modified robust variance estimator for generalized estimating equations with improved small-sample performance." \emph{Statistics in Medicine}, 30(11), 1278-1291. ISSN 02776715. doi:10.1002/sim.4150
}
\seealso{
\code{\link[geepack]{geeglm}}, \code{\link[geesmv]{GEE.var.pan}}, \code{\link[geesmv]{GEE.var.md}} and \code{\link[geesmv]{GEE.var.wl}}
}
\author{
Elizabeth Wynn, \pkg{geesmv}-authors for underlying code used in small sample size variance estimators.
}
