% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsigmag.R
\name{get_SigmaG.glmm_nb_mod}
\alias{get_SigmaG.glmm_nb_mod}
\title{Calculate Sigma and G matrices for Pseudo-likelihood Negative Binomial Mixed Model}
\usage{
\method{get_SigmaG}{glmm_nb_mod}(object, details = 0)
}
\arguments{
\item{object}{A \code{glmm_nb_mod} model object from the \code{glmm_nb_lmer} function}

\item{details}{If larger than 0 some timing details are printed.}
}
\value{
\item{Sigma}{The covariance matrix of Y}
\item{G}{The G matrices that sum up to Sigma}
\item{n.ggamma}{The number of G matrices (Called M in Kenward and Roger 1997)}
}
\description{
Calculates the Sigma and G components described in Kenward and
Roger (1997) while accounting for \code{glmm_nb_mod} object
weights. Function is used in calculating Kenward-Roger degrees of
freedom in functions from the \pkg{pbkrtest}-package such as \code{KRmodcomp}.
}
\references{
Kenward MG, Roger JH (1997). "Small Sample Inference for Fixed Effects from Restricted Maximum Likelihood." \emph{Biometrics}, 53(3), 983. ISSN 0006341X. doi: 10.2307/2533558.
}
\seealso{
\code{\link{glmm_nb_mod}}, \code{\link[pbkrtest]{get_SigmaG}}, \code{\link[pbkrtest]{KRmodcomp}}
}
\author{
Elizabeth Wynn and Camille Moore, underlying code drawn from code by \pkg{pbkrtest}-authors.
}
