% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_nb_lmer.R
\name{glmm_nb_lmer}
\alias{glmm_nb_lmer}
\title{Pseudo-Likelihood Negative Binomial Mixed Model}
\usage{
glmm_nb_lmer(
  formula,
  data,
  niter = 40,
  epsilon = 1e-08,
  verbose = FALSE,
  REML = TRUE
)
}
\arguments{
\item{formula}{A two-sided linear formula describing both the fixed-effects and random-effects parts of the model using the syntax of the lme4 package.}

\item{data}{An optional data frame containing the model variables.}

\item{niter}{Maximum number of iterations.}

\item{epsilon}{Positive convergence tolerance.}

\item{verbose}{Logical. Should the number of iterations and computational time be printed?}

\item{REML}{Logical. Should the models be fit with REML or regular ML?}
}
\value{
An object of class \code{\link{glmm_nb_mod}}
}
\description{
This function fits a negative binomial mixed model using a
psuedo-likelihood approach
}
\details{
#' This function is similar to the function \code{\link[NBZIMM]{glmm.nb}}
from the \pkg{NBZIMM} package, though it utilizes the \pkg{lmerTest} package
rather than \pkg{nlme} package and returns and object compatible with the \pkg{pkrtest} package
which allows for the calculation of Kenward-Roger degrees of freedom.
(see \code{\link[pbkrtest]{krmodcomp}},
\code{\link[lmerTest]{contest.lmerModLmerTest}})
}
\examples{
data("simdata")
sample_meta_data <- simdata$metadata

#Subset down to one observation (i.e. gene)
counts=simdata$counts[1,]

#Combine counts, metadata into dataframe
df=cbind(counts, sample_meta_data)

#Fit the Model
fit.glmm.nb <- glmm_nb_lmer(formula =counts ~ group * time + (1|ids),
                            data=df, REML = T)
}
\references{
Zhang X, Mallick H, Tang Z, Zhang L, Cui X, Benson AK, Yi N (2017). "Negative binomial mixed models for analyzing microbiome count data." \emph{BMC Bioinformatics}, 18(1), 4. ISSN 14712105. doi:10.1186/s12859-016-1441-7
}
\seealso{
\code{\link{glmm_nb_mod}}, \code{\link[lmerTest]{lmer}}
}
\author{
Elizabeth Wynn and Camille Moore, underlying code drawn from code by \pkg{NBZIMM}-authors.
}
