% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_nb_lmer.R
\docType{class}
\name{glmm_nb_mod-class}
\alias{glmm_nb_mod-class}
\alias{glmm_nb_mod}
\title{Class for Pseudo-Likelihood Negative Binomial Mixed Models}
\value{
An object of class \code{glmm_nb_mod} with slots as in
\code{\link[lmerTest]{lmerModLmerTest}} and a few
additional slots as described below.
}
\description{
Class for Pseudo-Likelihood Negative Binomial Mixed Models
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Final negative binomial model dispersion estimate}

\item{\code{niter}}{Number of iterations until model convergence (or iteration limit was reached)}
}}

\author{
Elizabeth Wynn
}
