% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSeq_summary.R
\name{corrSeq_summary}
\alias{corrSeq_summary}
\title{Function to summarize individual regression coefficients}
\usage{
corrSeq_summary(
  corrSeq_results = NULL,
  coefficient = NULL,
  p_adj_method = "BH",
  ddf = "Satterthwaite",
  sort_results = T
)
}
\arguments{
\item{corrSeq_results}{Results object from running \code{\link{corrSeq_fit}}.}

\item{coefficient}{Name of the coefficient being summarized.}

\item{p_adj_method}{Method for adjusting the raw p-values.}

\item{ddf}{Method for computing the degrees of freedom.}

\item{sort_results}{Should the results table be sorted by adjusted p-value?}

\item{summary_table}{A summary table including the gene name, estimate, standard error, degrees of freedom, test statistic, and raw and adjusted p-value.}

\item{singular_fits}{Gene names for genes that resulted in singular model fits. The summary information for these genes will be NA. Not applicable for models fit using \code{"gee"}.}

\item{method}{Method used to fit the models.}
}
\value{
This function returns a list object with the following components:
}
\description{
Conducts t-tests on individual regression coefficients from models fit using \code{\link{corrSeq_fit}}.
}
\examples{
data("simdata")
sample_meta_data <- simdata$metadata

## Subset down to 10 observation (i.e. gene)
counts=simdata$counts[1:10,]

## Fit NBMM-PL models
nbmm_pl_fit <- corrSeq_fit(formula = ~ group * time+(1|ids),
                          expr_mat = counts,
                          sample_data = sample_meta_data,
                          method="nbmm_pl")

## Summarize the group coefficient with Satterthwaite degrees of freedom
model_sum <-corrSeq_summary(corrSeq_results = nbmm_pl_fit,
                             coefficient = "group",
                             p_adj_method = 'BH',
                             ddf = 'Satterthwaite',
                             sort_results = T)


}
\seealso{
\code{\link{corrSeq_fit}} \code{\link{geeglm_small_samp}}, \code{\link{glmm_nb_lmer}}, \code{\link[lmerTest]{lmer}}, \code{\link[glmmADMB]{glmmadmb}}
}
\author{
Elizabeth Wynn
}
